server/pyproject.toml
[tool.poetry]
name = "newgeotime"
version = "0.1.0"
description = ""
authors = ["ileschenko <ileschenko@ptsecurity.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic-settings = "^2.2.1"
aio-pika = "^9.4.0"
pika = "^1.3.2"
loguru = "^0.7.2"
celery = "^5.3.6"
opencv-python = "^4.9.0.80"
sqlalchemy = "^2.0.27"
ffmpeg = "^1.4"
websockets = "^12.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
orjson = "^3.9.15"
redis = "^5.0.1"
pandas = "^2.2.1"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
